% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{blockDiag}
\alias{blockDiag}
\title{Compute a block diagonal matrix w/ constant blocks}
\usage{
blockDiag(Amat, nrep)
}
\arguments{
\item{Amat}{matrix to populate the diagaonal blocks}

\item{nrep}{number of blocks on the diagonal}
}
\description{
The function returns kronecker(diag(nrep), Amat), but is computed more efficiently
}
