% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{computeTimeRemaining}
\alias{computeTimeRemaining}
\title{Estimate the remaining time in the MCMC based on previous samples}
\usage{
computeTimeRemaining(nsi, timer0, nsims, nrep = 100)
}
\arguments{
\item{nsi}{Current iteration}

\item{timer0}{Initial timer value, returned from \code{proc.time()[3]}}

\item{nsims}{Total number of simulations}

\item{nrep}{Print the estimated time remaining every \code{nrep} iterations}
}
\value{
Table of summary statistics using the function \code{summary}
}
\description{
Estimate the remaining time in the MCMC based on previous samples
}
