% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{f_ns}
\alias{f_ns}
\title{Compute the Nelson-Siegel Curves}
\usage{
f_ns(tau, lambda_p = 0.0609)
}
\arguments{
\item{tau}{the \code{m x 1} vector of observation points (in months)}

\item{lambda}{the nonlinear parameter}
}
\value{
The \code{m x 3} matrix of Nelson-Siegel curves
evaluated at \code{tau}
}
\description{
Compute the three Nelson-Siegel curves at a given vector of points
}
\examples{
tau = 1:300
F_ns = f_ns(tau)
plot(tau, F_ns[,1], ylim = range(F_ns), type='l', lwd=4)
for(j in 2:3) lines(tau, F_ns[,j], lwd=4,lty=j)

}
