% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{fdlm}
\alias{fdlm}
\title{MCMC Sampling Algorithm for the FDLM}
\usage{
fdlm(Y, tau, K = NULL, nsims = 10000, burnin = 1000,
  mcmc_params = list("beta", "fk"), useFastImpute = TRUE)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{tau}{vector of observation points (\code{m}-dimensional)}

\item{K}{the number of factors; if NULL, use SVD-based proportion of variability explained}

\item{nsims}{number of MCMC iterations to record}

\item{burnin}{length of the burnin, which is discarded}

\item{mcmc_params}{named list of parameters for which we store the MCMC output;
must be one or more of "beta" (factors),  "fk" (FLCs), "sigma_e" (observation error SD), "Wt" (evolution error variance), "Yhat" (fitted values)}

\item{useFastImpute}{logical; when TRUE, use imputation/projection scheme for the dynamic factors; otherwise use full state space model for factors (slower)}
}
\value{
A named list of the \code{nsims} MCMC samples for the parameters named in \code{mcmc_params}
}
\description{
Runs the MCMC for the (univariate) FDLM under some default conditions:
1) A random walk model for the dynamic factors;
2) A full \code{K x K} non-dynamic evolution error variance matrix;
3) A non-dynamic scalar multiple of the identity for the observation error variance.
}
\note{
If \code{Tm} is large, then storing all posterior samples for \code{Yhat}, which is \code{nsims x T x m},  may be inefficient
}
\examples{
# Read in the yield curve data:
data("US_Yields")

# Restrict to dates since 2006:
Y = Y[which(dates > as.Date("2006-01-01")),];

# Run the MCMC:
mcmc_output = fdlm(Y, tau, K = 3, nsims = 1000)
}
