% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{fdlm_factor}
\alias{fdlm_factor}
\title{Sample the dynamic factors}
\usage{
fdlm_factor(Y, sigma_et, Wt, Fmat = NULL, YF = NULL, Gt = NULL,
  kfas_model = NULL, useFastImpute = FALSE)
}
\arguments{
\item{Y}{the \code{T x m} data observation matrix, where \code{T} is the number of time points and \code{m} is the number of observation points (\code{NA}s allowed)}

\item{sigma_et}{\code{T}- or \code{1}-dimensional vector of observation error standard deviation(s)}

\item{Wt}{\code{K x K} matrix or \code{K x K x T} array of evolution error covariances}

\item{Fmat}{\code{m x K} matrix of FLCs; only needed for \code{useFastImpute = FALSE}}

\item{YF}{\code{T x K} matrix of data \code{Y} projected onto FLCs, \code{Y\%*\%Fmat}, which takes the place
of \code{Y}; only needed for \code{useFastImpute = TRUE}}

\item{Gt}{\code{K x K} evolution matrix; if NULL, set as identity (for random walk)}

\item{kfas_model}{\code{SSModel} object from KFAS package; if NULL, construct model w/in the sampler (might be slower!)}

\item{useFastImpute}{logical; when TRUE, use imputation/projection scheme for the dynamic factors; otherwise use full state space model for factors (slower)}
}
\value{
The \code{T x K} matrix of dynamic factors, \code{Beta}.
}
\description{
Sample the dynamic factors (latent state space variables) using the simulation smoothing
algorithm of Koopman and Durbin (2001), implemented in the KFAS package.
}
\note{
The sampler has two options: \code{useFastImpute = TRUE}, in which the response is
\code{YF = Y\%*\%Fmat} (\code{T x K}) and the observation matrix is the identity (\code{K x K});
and \code{useFastImpute = FALSE}, in which the response is \code{Y}  (\code{T x m})
and the observation matrix is \code{Fmat} (\code{m x K}). Recall that typically \code{K < < m},
so \code{useFastImpute = TRUE} is often much faster.
}
\examples{
# Read in the yield curve data:
data("US_Yields")

# Restrict to dates since 2006:
Y = Y[which(dates > as.Date("2006-01-01")),];

# This is a simple (yet meaningless) example:
K = 3
Beta = fdlm_factor(Y, sigma_et = 1, Wt = diag(K), Fmat = diag(ncol(Y))[,1:K])
}
