% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{fdlm_flc}
\alias{fdlm_flc}
\title{Factor Loading Curve Sampling Algorithm}
\usage{
fdlm_flc(BtY, Beta, Psi, BtB, Omega, BtCon = NULL, lambda, sigmat2)
}
\arguments{
\item{BtY}{\code{J x T} matrix \code{B.t()*Y} for basis matrix B}

\item{Beta}{\code{T x K} matrix of factors}

\item{Psi}{\code{J x K} matrix of previous factor loading curve coefficients}

\item{BtB}{\code{J x J} matrix of \code{B.t()*B}}

\item{Omega}{\code{J x J} prior precision/penalty matrix}

\item{BtCon}{(optional) \code{J x Jc} matrix of additional constraints, pre-multiplied by B.t()}

\item{lambda}{\code{K}-dimensional vector of prior precisions}

\item{sigmat2}{\code{T}-dimensional vector of time-dependent observation error variances}
}
\value{
Psi \code{J x K} matrix of factor loading curve coefficients
}
\description{
Sample the factor loading curve basis coefficients subject to an orthonormality constraint.
Additional linear constraints may be included as well.
}
\note{
This is a wrapper for Rcpp functions for the special cases of
\code{K = 1} and whether or not additional (linear) constraints are included,
i.e., whether or not \code{BtCon} is non-\code{NULL}.
}
