% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getSplineInfo_d}
\alias{getSplineInfo_d}
\title{Construct the spline basis and penalty matrices}
\usage{
getSplineInfo_d(tau, num.knots = NULL, orthonormalize = TRUE)
}
\arguments{
\item{tau}{\code{m x d} matrix of coordinates, where \code{m} is the number of observation points and \code{d} is the dimension}

\item{num.knots}{the number of knots to include}

\item{orthonormalize}{logical; if TRUE, orthornomalize the basis matrix}
}
\description{
Given input points in \code{d} dimensions, construct a low-rank thin plate spline basis matrix
and penalty matrix.
}
\note{
The knot locations are selected using a space-filling algorithm.
}
