% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSV_source.R
\name{initCommonSV}
\alias{initCommonSV}
\title{Initialize the (common) SV parameters}
\usage{
initCommonSV(omega)
}
\arguments{
\item{omega}{\code{T x m} matrix of residuals}
}
\value{
List of relevant components:
\itemize{
\item the \code{T x 1} error standard deviations \code{sigma_et},
\item the \code{T x 1} log-volatility \code{ht},
\item the log-vol unconditional mean \code{h_mu},
\item the log-vol AR(1) coefficient \code{h_phi},
\item the log-vol innovation standard deviation \code{h_sigma_eta}
}
}
\description{
Compute initial values for common stochastic volatility  parameters
}
