% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot a curve given posterior samples}
\usage{
plot_curve(post_f, tau = NULL, alpha = 0.05, include_joint = TRUE)
}
\arguments{
\item{post_f}{\code{Ns x m} matrix of \code{Ns} posterior simulations
of the curve at \code{m} points}

\item{tau}{\code{m x 1} vector of observation points}

\item{alpha}{confidence level for the bands}

\item{include_joint}{logical; if TRUE, include joint bands (as well as pointwise)}
}
\description{
Plot the posterior mean, simultaneous and pointwise 95\% credible bands
for a curve given draws from the posterior distribution
}
