% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot_factors}
\alias{plot_factors}
\title{Plot the factors}
\usage{
plot_factors(post_beta, dates = NULL)
}
\arguments{
\item{post_beta}{the \code{Nsims x T x K} array of \code{Nsims} draws from the posterior
distribution of the \code{T x K} matrix of factors, \code{beta}}

\item{dates}{\code{T x 1} vector of dates or labels corresponding to the time points}
}
\description{
Plot posterior mean of the factors together with the simultaneous and pointwise
95\% credible bands.
}
