% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sample_lambda}
\alias{sample_lambda}
\title{Factor loading curve smoothing parameter sampler}
\usage{
sample_lambda(lambda, Psi, Omega = NULL, uniformPrior = TRUE,
  orderLambdas = TRUE)
}
\arguments{
\item{lambda}{\code{K}-dimensional vector of smoothing parameters (prior precisions)
from previous MCMC iteration}

\item{Psi}{\code{J x K} matrix of basis coefficients, where \code{J} is the number of
basis functions and \code{K} is the number of factors}

\item{Omega}{\code{J x J} penalty matrix; if NULL, assume it is diag(0, 0, 1,...,1)}

\item{uniformPrior}{logical; when TRUE, use a uniform prior on prior standard deviations,
\code{1/sqrt{lambda[k]}}; otherwise use independent Gamma(0.001, 0.001) prior for each \code{lambda[k]}}

\item{orderLambdas}{logical; when TRUE, enforce the ordering constraint \code{lambda[1] > ... > lambda[K]}
for identifiability}
}
\value{
The \code{K}-dimensional vector of samoothing parameters, \code{lambda}.
}
\description{
Sample the smoothing parameters for each factor loading curve.
}
