% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sample_Wt}
\alias{sample_Wt}
\title{Evolution error variance sampler}
\usage{
sample_Wt(resBeta, Rinv = diag(1, K), rh0 = K, useDiagonal = FALSE)
}
\arguments{
\item{resBeta}{\code{(T-1) x K} matrix of residuals from the evolution equation}

\item{Rinv}{\code{K x K}; only needed for \code{useDiagonal = FALSE}}

\item{rh0}{scalar prior scale parameter; only needed for \code{useDiagonal = FALSE}}

\item{useDiagonal}{logical; when TRUE, use a diagonal covariance matrix; otherwise, use a full matrix}
}
\value{
The \code{K x K} evolution error covariance matrix, \code{Wt}.
}
\description{
Sample the (non-dynamic) evolution error variance matrix.
}
\note{
The sampler has two options: \code{useDiagonal = TRUE}, in which the evolution
error variance matrix is diagional with independent Gamma(0.001,0.001) priors on the
\code{K} diagonal component precisions; and \code{useDiagonal = FALSE}, in which the
\code{K x K} evolution error variance matrix is full with an inverse Wishart prior
with prior precision \code{Rinv} and scale parameter \code{rh0}. The default is
\code{Rinv = diag(1, K)} and \code{rh0 = K}.
}
\examples{
# Simulate an example:
n = 100; K = 3
resBeta = matrix(rnorm(n*K), nr = n, nc = K)

# Sample the variance:
sample_Wt(resBeta)
}
