% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sampleARphi}
\alias{sampleARphi}
\title{Sample the autoregressive coefficients in an AR(1) Model}
\usage{
sampleARphi(yt, phi_j, sigma_tj, prior_phi = NULL)
}
\arguments{
\item{yt}{the \code{T x p} matrix of centered multivariate time series
(i.e., the time series minus the unconditional means, \code{mu})}

\item{phi_j}{the \code{p x 1} vector of previous AR(1) coefficients}

\item{sigma_tj}{the \code{(T-1) x p} matrix or \code{p x 1} vector of error standard deviations}

\item{prior_phi}{the parameters of the prior for the AR(1) coefficients \code{phi_j};
either \code{NULL} for uniform on [-0.99,0.99] or a 2-dimensional vector of (shape1, shape2) for a Beta prior
on \code{[(phi_j + 1)/2]}}
}
\value{
\code{p x 1} vector of sampled AR(1) coefficient(s)
}
\description{
Compue one draw of the autoregressive coefficients \code{phi} in an AR(1) model.
The sampler also applies to a multivariate case with independent components.
}
\details{
Sample the AR(1) coefficients \code{phi_j} using the model

\code{y_tj = mu_j + phi_j(y_{t-1,j} - mu_j) + e_tj},

with \code{e_tj ~ N(0, sigma[j]^2)}
}
\note{
For the standard AR(1) case, \code{p = 1}. However, the function applies more
generally for sampling \code{p > 1} independent AR(1) processes (jointly).
}
