% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSV_source.R
\name{sampleCommonLogVols}
\alias{sampleCommonLogVols}
\title{Sample the latent log-volatilities, common to m-dimensional time series}
\usage{
sampleCommonLogVols(h_y, h_prev, h_mu, h_phi, h_sigma_eta)
}
\arguments{
\item{h_y}{the \code{T x m} matrix of data, which follows one join SV model}

\item{h_prev}{the \code{T x 1} matrix of the previous log-vols}

\item{h_mu}{the log-vol unconditional mean}

\item{h_phi}{the log-vol AR(1) coefficient}

\item{h_sigma_eta}{the log-vol innovation standard deviation}
}
\value{
\code{T x 1} matrix of simulated log-vols
}
\description{
Compute one draw of the log-volatilities using a discrete mixture of Gaussians
approximation to the likelihood (see Omori, Chib, Shephard, and Nakajima, 2007)
where the log-vols are assumed to follow an AR(1) model. The model assumes that
the volatility are common to an m-dimensional time series.
}
