% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSV_source.R
\name{sampleCommonSV}
\alias{sampleCommonSV}
\title{Sample the (common) stochastic volatility parameters}
\usage{
sampleCommonSV(omega, svParams, prior_phi = c(20, 1.5))
}
\arguments{
\item{omega}{\code{T x m} matrix of residuals}

\item{svParams}{list of parameters to be updated (see Value below)}

\item{prior_phi}{the parameters of the prior for the log-volatilty AR(1) coefficient \code{h_phi};
either \code{NULL} for uniform on [-1,1] or a 2-dimensional vector of (shape1, shape2) for a Beta prior
on \code{[(h_phi + 1)/2]}}
}
\value{
List of relevant components:
\itemize{
\item the \code{T x 1} error standard deviations \code{sigma_et},
\item the \code{T x 1} log-volatility \code{ht},
\item the log-vol unconditional mean \code{h_mu},
\item the log-vol AR(1) coefficient \code{h_phi},
\item the log-vol innovation standard deviation \code{h_sigma_eta}
}
}
\description{
Compute one draw for each of the parameters in the stochastic volatility model
where each component of omega_{j,t} has a common dynamic variance.
}
