% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleFLC_1}
\alias{sampleFLC_1}
\title{Factor Loading Curve Sampling Algorithm for K=1}
\usage{
sampleFLC_1(BtY, Beta, Psi, BtB, Omega, lambda, sigmat2)
}
\arguments{
\item{BtY}{\code{J x T} matrix \code{B.t()*Y} for basis matrix B}

\item{Beta}{\code{T x 1} matrix of factors}

\item{Psi}{\code{J x 1} matrix of previous factor loading curve coefficients}

\item{BtB}{\code{J x J} matrix of \code{B.t()*B}}

\item{Omega}{\code{J x J} prior precision/penalty matrix}

\item{lambda}{\code{1}-dimensional vector of prior precisions}

\item{sigmat2}{\code{T}-dimensional vector of time-dependent observation error variances}
}
\value{
Psi \code{J x 1} matrix of factor loading curve coefficients
}
\description{
Sample the factor loading curve basis coefficients for K=1 factor.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
