% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleFLC_orthog}
\alias{sampleFLC_orthog}
\title{Factor Loading Curve Sampling Algorithm}
\usage{
sampleFLC_orthog(BtY, Beta, Psi, Omega, lambda, sigmat2)
}
\arguments{
\item{BtY}{\code{J x T} matrix \code{B.t()*Y} for basis matrix B}

\item{Beta}{\code{T x K} matrix of factors}

\item{Psi}{\code{J x K} matrix of previous factor loading curve coefficients}

\item{Omega}{\code{J x J} prior precision/penalty matrix}

\item{lambda}{\code{K}-dimensional vector of prior precisions}

\item{sigmat2}{\code{T}-dimensional vector of time-dependent observation error variances}
}
\value{
Psi \code{J x K} matrix of (orthogonal) factor loading curve coefficients
}
\description{
Sample the factor loading curve basis coefficients subject to
an orthonormality constraint for the special case in which \code{BtB = diag(J)}.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
