% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sampleVAR}
\alias{sampleVAR}
\title{Sample the VAR coefficient matrix}
\usage{
sampleVAR(ytc, Sigma, priorPrec = NULL, stationary = TRUE)
}
\arguments{
\item{ytc}{the \code{T x p} matrix of (centered) multivariate time series}

\item{Sigma}{the \code{p x p} error variance matrix}

\item{priorPrec}{the \code{p x p} prior precision matrix;
if \code{NULL}, use \code{diag(10^-6, p)}}

\item{stationary}{logical; if \code{TRUE}, resample until stationary}
}
\value{
The \code{p x p} VAR coefficient matrix.
}
\description{
Compue one draw of the VAR coefficient matrix \code{G} assuming a
Gaussian prior (with mean zero). The sampler also assumes the VAR
is lag-1.
}
