% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_model}
\alias{update_model}
\title{Update the KFAS model}
\usage{
update_model(pars, model, tau, f_p, orthogonalize)
}
\arguments{
\item{pars}{vector of parameters}

\item{model}{object of class SSModel() which should be updated}

\item{tau}{vector of observation points (\code{m}-dimensional)}

\item{f_p}{a function to compute the parametric component, which must return a \code{m x K_p} matrix
for \code{K_p} the number of parametric curves; may include a (scalar) nonlinear parameter argument}

\item{orthogonalize}{logical; when TRUE, orthogonalize the loading curve matrix}
}
\value{
The updated \code{model}
}
\description{
Update the KFAS model using the given parameters, assuming a specific form.
}
\note{
This function is necessary for state space optimization (initialization)
}
