% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_samplers.R
\name{sampleARmu}
\alias{sampleARmu}
\title{Sample the unconditional mean in a VAR}
\usage{
sampleARmu(yt, G, Sigma, priorPrec = NULL)
}
\arguments{
\item{yt}{the \code{T x p} matrix of multivariate time series}

\item{G}{the \code{p x p} VAR coefficient matrix}

\item{Sigma}{the \code{p x p} error variance matrix}

\item{priorPrec}{the \code{p x p} prior precision matrix;
if \code{NULL}, use \code{diag(10^-6, p)}}
}
\value{
The \code{p x 1} matrix of unconditional means.
}
\description{
Compue one draw of the unconditional mean \code{mu} in a VAR assuming a
Gaussian prior (with mean zero).
}
\details{
Sample the unconditional mean \code{mu} using the model

\code{y_t = mu + G(y_{t-1} - mu) + e_t},

with \code{e_t ~ N(0, Sigma)} and prior \code{mu ~ N(0, solve(priorPrec))}
}
