% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleFLC}
\alias{sampleFLC}
\title{Sample the Factor Loading Curves}
\usage{
sampleFLC(BtY, Beta, Psi, BtB, Omega, lambda, sigmat2)
}
\arguments{
\item{BtY}{(J x T) matrix B.t()*Y for basis matrix B}

\item{Beta}{(T x K) matrix of factors}

\item{Psi}{(J x K) matrix of previous factor loading curve coefficients}

\item{BtB}{(J x J) matrix of B.t()*B}

\item{Omega}{(J x J) prior precision/penalty matrix}

\item{sigmat2}{T-dimensional vector of time-dependent observation error variances}

\item{lamda}{K-dimensional vector of prior precisions}
}
\description{
Sample the Factor Loading Curves
}
