% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleFLC_orthog}
\alias{sampleFLC_orthog}
\title{Sample the Factor Loading Curves, For the special case in which BtB = diag(J)}
\usage{
sampleFLC_orthog(BtY, Beta, Psi, Omega, lambda, sigmat2)
}
\arguments{
\item{BtY}{(J x T) matrix B.t()*Y for basis matrix B}

\item{Beta}{(T x K) matrix of factors}

\item{Psi}{(J x K) matrix of previous factor loading curve coefficients}

\item{Omega}{(J x J) prior precision/penalty matrix}

\item{sigmat2}{T-dim vector of time-dependent observation error variances}

\item{lamda}{K-dimensional vector of prior precisions}
}
\description{
Sample the Factor Loading Curves, For the special case in which BtB = diag(J)
}
